minor fixes

+++ structures/Block.ts
@@ -13,6 +13,7 @@ export class Block {
         const match = await sql.fetchColumn("blocks", "block match")
         if (String(match) === "exact") {
             if (words.some((w: string) => w.includes(message.content) || (asterisk ? message.content.match(/\*/g) : false))) {
+                if (!message.guild?.me?.permissions.has("MANAGE_MESSAGES")) return message.channel.send("I need the **Manage Messages** permission in order to delete the message with the blocked word.")
                 const reply = await message.reply("Your post was removed because it contained a blocked word.")
                 await message.delete()
                 reply.delete({timeout: 10000})
@@ -20,6 +21,7 @@ export class Block {
         } else {
             for (let i = 0; i < words.length; i++) {
                 if (message.content.includes(words[i]) || (asterisk ? message.content.match(/\*/g) : false)) {
+                    if (!message.guild?.me?.permissions.has("MANAGE_MESSAGES")) return message.channel.send("I need the **Manage Messages** permission in order to delete the message with the blocked word.")
                     const reply = await message.reply("Your post was removed because it contained a blocked word.")
                     await message.delete()
                     reply.delete({timeout: 10000})
